Split context and config types across host layers
src/core/types.ts is now just the substrate: CoreContext (bus,
handlers, lifecycle, storage) and CoreConfig (extensions, backend).
Each host layers in its own surface:

- src/agent/host-types.ts defines AgentSurface and AgentConfigSurface
  with the LLM port, tool/instruction/skill registration, context
  producers, and provider config. AgentContext = CoreContext &
  AgentSurface; AgentConfig = CoreConfig & AgentConfigSurface.

- src/shell/host-types.ts defines ShellSurface and ShellConfigSurface
  with the compositor, remote sessions, input modes, slash commands,
  stream transforms, and palette overrides. ShellContext extends
  AgentContext; ShellConfig extends AgentConfig.

ExtensionContext = ShellContext and AgentShellConfig = ShellConfig
remain as back-compat aliases in the shell host. Internal imports
updated to pull from the owning host file. agent-sh/types subpath
now points at the main barrel so downstream consumers like ashi
keep resolving ExtensionContext.
