+++ src/agent/providers/openai-compatible.ts
@@ -3,7 +3,7 @@
  * Studio, vLLM, llama.cpp, …). No reasoning hook — the right shape depends
  * on which model the server is serving; user extensions can add one.
  */
-import type { ExtensionContext } from "../../core/types.js";
+import type { ExtensionContext } from "../../shell/host-types.js";
 
 export default function activate(ctx: ExtensionContext): void {
   const baseURL = process.env.OPENAI_BASE_URL;
