+++ src/extensions/file-autocomplete.ts
@@ -7,7 +7,7 @@
  */
 import * as fs from "node:fs";
 import * as path from "node:path";
-import type { ExtensionContext } from "../core/types.js";
+import type { ExtensionContext } from "../shell/host-types.js";
 
 export default function activate(ctx: ExtensionContext): void {
   ctx.bus.onPipe("autocomplete:request", (payload) => {
