+++ src/extensions/slash-commands.ts
@@ -11,7 +11,7 @@
  * and check payload.command / payload.commandArgs to add completions for any command.
  */
 import { palette as p } from "../utils/palette.js";
-import type { ExtensionContext } from "../core/types.js";
+import type { ExtensionContext } from "../shell/host-types.js";
 import { discoverSkills, loadSkillContent, type Skill } from "../agent/skills.js";
 import { reloadExtensions } from "../core/extension-loader.js";
